+++ CHANGELOG.md
@@ -14,6 +14,7 @@ This file is used to list changes made in each version of grafana.
 - Added support for custom plugin URLs
 - Added support for creating dashboards from template file
 - Fixes bug that didn't create folder when using action update and it did not exist yet
+- Fixes bug that didn't create datasource when using action update and it did not exist yet
 - Fixes bug where installing a lot of plugins would cause grafana systemd service to fail to start
 - Removes unecessary restart on plugin install